src/financial_knowledge/pyproject.toml
[project]
name = "financial-knowledge"
version = "0.1.0"
description = "Financial Knowledge Generator for Discord Bot"
dependencies = [
    "bot-common",
    "discord-webhook>=1.3.0",
    "python-dotenv>=1.0.0",
]

[tool.uv.sources]
bot-common = { workspace = true }
